Plus de global inutile
