+++ PHP/Controleur/ScenarioProf/evenement.php
@@ -84,7 +84,7 @@ if (@$_COOKIE['reload']==1){
 include("../../View/BarreHTML/BarreScenario.php");
 
 require('../../Modele/BDD/ConnectionBDD.php');
-require('FonctionScenario.php');
+require('../../Modele/FonctionScenario.php');
 // affichage des événements existants
 $pdo = ConnectionBDD::getInstance();
 $bdd = $pdo::getpdo();
