+++ PHP/Modele/Prof/modeleAfficheScenario2.php
@@ -1,0 +1,9 @@
+<?php
+function affpatient($bdd, $id)
+{
+    $sql = $bdd->prepare("SELECT * from patient where idpatient=?");
+    $sql->execute(array($id));
+    $array = $sql->fetch();
+    return $array;
+
+}
